Set `<leader>ct` shortcut to use Git ctags hook
In cbdcbce01dea1ab3850be2311f33f00d75f6088b we introduced a Tim
Pope-inspired Git template scheme that automatically ran `ctags`
asynchronously through a hook after certain source control operations.
This hook is bypassed if a user uses `<leader>ct` to reindex their tags
from within Vim.

This change binds `<leader>ct` to use a Vimscript function that will
identify if the `ctags` hook exists in the local repository and executes
it. The function falls back to existing functionality if the hook does
not exist.

This file is being out of `vimrc` as a first step toward breaking our
Vim configuration up by topic area.

Move ctags re-indexing to `vim/plugin`

As suggested by @gylaz and @teoljungberg in #455, this function and
mapping can be extracted to a separate file that vim with automatically
source.

what